Refactor `format` command to a standalone script
- Move `format` script into `contrib` directory
- Move contrib section of README to separate README within `contrib` directory
- Add section to contrib README describing the `format` script
