--- rdm/init_files/data/packages.yml
@@ -1,60 +1,0 @@
-accountsservice:
-  states: ii
-  version: 0.6.55-0ubuntu12~20.04.4
-acl:
-  states: ii
-  version: 2.2.53-6
-acpi-support:
-  states: ii
-  version: '0.143'
-acpid:
-  states: ii
-  version: 1:2.0.32-1ubuntu1
-adduser:
-  states: ii
-  version: 3.118ubuntu2
-adwaita-icon-theme:
-  states: ii
-  version: 3.36.1-2ubuntu0.20.04.2
-aisleriot:
-  states: ii
-  version: 1:3.22.9-1
-alsa-base:
-  states: ii
-  version: 1.0.25+dfsg-0ubuntu5
-alsa-topology-conf:
-  states: ii
-  version: 1.2.2-1
-alsa-ucm-conf:
-  states: ii
-  version: 1.2.2-1ubuntu0.7
-alsa-utils:
-  states: ii
-  version: 1.2.2-1ubuntu2.1
-anacron:
-  states: ii
-  version: 2.3-29
-apg:
-  states: ii
-  version: 2.2.3.dfsg.1-5
-app-install-data-partner:
-  states: ii
-  version: '19.04'
-apparmor:
-  states: ii
-  version: 2.13.3-7ubuntu5.1
-apport:
-  states: ii
-  version: 2.20.11-0ubuntu27.17
-apport-gtk:
-  states: ii
-  version: 2.20.11-0ubuntu27.17
-apport-symptoms:
-  states: ii
-  version: '0.23'
-appstream:
-  states: ii
-  version: 0.12.10-2
-apt:
-  states: ii
-  version: 2.0.5
